[PR116778][LRA]: Check pseudos assigned to FP after rematerialization to build live ranges
This is a better fix of the PR permitting to avoid building live
ranges after rematerialization.  It checks that FP can not be
eliminated now and that pseudos assigned to FP will be spilled.  In
this case we need to build live ranges after rematerialization for
correct assignments of stack slots to spilled pseudos involved in
rematerialization.

gcc/ChangeLog:

	PR rtl-optimization/116778
	* ira-int.h (x_ira_class_hard_reg_index): Fix comment typo.
	* lra-eliminations.cc (lra_fp_pseudo_p): New function.
	* lra-int.h (lra_fp_pseudo_p): External declaration.
	* lra-spills.cc (lra_need_for_spills_p): Fix formatting.
	* lra.cc (lra): Use lra_fp_pseudo_p in lra_create_live_range after
	lra_remat.