LoongArch: Implement sge and sgeu.
The original implementation of the function loongarch_extend_comparands
only prevented op1 from being loaded into the register when op1 was
const0_rtx.  It has now been modified so that op1 is not loaded into
the register as long as op1 is an immediate value.  This allows
slt{u}i to be generated instead of slt{u} if the conditions are met.

gcc/ChangeLog:

	* config/loongarch/loongarch.cc
	(loongarch_canonicalize_int_order_test): Support GT GTU LT
	and LTU.
	(loongarch_extend_comparands): Expand the scope of op1 from
	0 to all immediate values.
	* config/loongarch/loongarch.md
	(*sge<u>_<X:mode><GPR:mode>): New template.

gcc/testsuite/ChangeLog:

	* gcc.target/loongarch/sign-extend-3.c: New test.