RISC-V: Emit vector shift pattern for const_vector [PR117353].
In PR117353 and PR117878 we expand a const vector during reload.  For
this we use an unpredicated left shift.  Normally an insn like this is
split but as we introduce it late and cannot create pseudos anymore
it remains unpredicated and is not recognized by the vsetvl pass (where
we expect all insns to be in predicated RVV format).

This patch directly emits a predicated shift instead.  We could
distinguish between !lra_in_progress and lra_in_progress and emit
an unpredicated shift in the former case but we're not very likely
to optimize it anyway so it doesn't seem worth it.

	PR target/117353
	PR target/117878

gcc/ChangeLog:

	* config/riscv/riscv-v.cc (expand_const_vector): Use predicated
	instead of simple shift.

gcc/testsuite/ChangeLog:

	* gcc.target/riscv/rvv/autovec/pr117353.c: New test.