ada: Correct xref of operator expression function body
For an expression function body that is an operator,
make sure the xref entry in the ALI file points one past the
double quote mark. For example, if the name is ">", point
to the greater-than symbol, not the double quote.
This was already the case for proper bodies.

gcc/ada/ChangeLog:

	* sem_util.adb (Copy_Subprogram_Spec): When copying
	an N_Defining_Operator_Symbol, create an
	N_Defining_Operator_Symbol rather than an N_Defining_Identifier.
	This will cause Lib.Xref to "skip the initial quote for
	navigation purposes" (as per comments in Lib.Xref body).