avoid-store-forwarding: fix reg init on load-eliminiation [PR117835]
During the initialization of the base register for the zero-offset
store, in the case that we are eliminating the load, we used a
paradoxical subreg assuming that we don't care about the higher bits
of the register. This led to writing wrong values when we were not
updating the whole register.

This patch fixes the issue by zero-extending the value stored in the
base register instead of using a paradoxical subreg.

Bootstrapped/regtested on x86 and AArch64.

	PR rtl-optimization/117835
	PR rtl-optimization/117872

gcc/ChangeLog:

	* avoid-store-forwarding.cc
	(store_forwarding_analyzer::process_store_forwarding):
	Zero-extend the value stored in the base register instead of
	using a paradoxical subreg.

gcc/testsuite/ChangeLog:

	* gcc.target/i386/pr117835.c: New test.