c++: Handle attributes on exception declarations [PR110345]
This is a continuation of the series for the ignorability of standard
attributes.

I've added a test for assume attribute diagnostics appertaining to various
entities (mostly invalid) and while doing that, I've discovered that
attributes on exception declarations were mostly ignored, this patch
adds the missing cp_decl_attributes call and also in the
cp_parser_type_specifier_seq case differentiates between attributes and
std_attributes to be able to differentiate between attributes which apply
to the declaration using type-specifier-seq and attributes after the type
specifiers.

2024-12-18  Jakub Jelinek  <jakub@redhat.com>

	PR c++/110345
	* parser.cc (cp_parser_type_specifier_seq): Chain cxx11_attribute_p
	attributes after any type specifier in the is_declaration case
	to std_attributes rather than attributes.  Set also ds_attribute
	or ds_std_attribute locations if not yet set.
	(cp_parser_exception_declaration): Pass &type_specifiers.attributes
	instead of NULL as last argument, call cp_decl_attributes.

	* g++.dg/cpp0x/attr-assume1.C: New test.