noncontiguous ifcombine: skip marking of non-SSA_NAMEs [PR117915]
When ifcombine_mark_ssa_name is called directly, rather than by
ifcombine_mark_ssa_name_walk, we need to check that name is an
SSA_NAME at the caller or in the function itself.  For convenience and
safety, I'm moving the checks from _walk to the implementation proper.


for  gcc/ChangeLog

	PR tree-optimization/117915
	* tree-ssa-ifcombine.cc (ifcombine_mark_ssa_name): Move
	preconditions from...
	(ifcombine_mark_ssa_name_walk): ... here.

for  gcc/testsuite/ChangeLog

	PR tree-optimization/117915
	* gcc.dg/pr117915.c: New.